+++ packages/iov-core/types/index.d.ts
@@ -1,6 +1,6 @@
 export { Address, ChainId, Nonce, SendTransaction, TokenTicker } from "@iov/bcp";
 export { Ed25519HdWallet, Ed25519Wallet, HdPaths, Keyring, Secp256k1HdWallet, UserProfile, Wallet, WalletId, WalletImplementationIdString, WalletSerializationString, } from "@iov/keycontrol";
-export { JsonRpcSigningServer } from "./jsonrpcsigningserver";
 export { JsRpcClient, jsRpcCode, JsRpcErrorResponse, JsRpcRequest, JsRpcResponse, JsRpcSuccessResponse, parseJsRpcErrorResponse, parseJsRpcId, parseJsRpcRequest, parseJsRpcResponse, } from "./jsrpc";
+export { JsRpcSigningServer } from "./jsrpcsigningserver";
 export { MultiChainSigner } from "./multichainsigner";
 export { GetIdentitiesAuthorization, SignAndPostAuthorization, SigningServerCore } from "./signingservercore";